pyproject.toml
[project]
name = "ptylab"
version = "0.2.1"
description = "A cross-platform, open-source inverse modeling toolbox for conventional and Fourier ptychography"
authors = [
    { name = "Lars Loetgering", email = "lars.loetgering@fulbrightmail.org" },
    { name = "PtyLab Team" },
]
readme = "README.md"
requires-python = ">=3.9, <3.13"
dependencies = [
    "numpy>=1.22",
    "matplotlib~=3.7",
    "h5py~=3.9",
    "scipy~=1.11",
    "scikit-image>=0.21.0",
    "scikit-learn~=1.3",
    "tqdm~=4.65",
    "pyqtgraph~=0.13",
    "tables~=3.8",
    "bokeh~=3.2",
    "tensorflow~=2.14",
    "PyQt5~=5.15",
]

[project.optional-dependencies]
dev = [
    "black~=23.7",
    "ipykernel~=6.25",
    "pre-commit~=4.0",
]
cuda12 = ["cupy-cuda12x"]
cuda13 = ["cupy-cuda13x"]

[project.scripts]
ptylab = "PtyLab.utils.cli:main"

[build-system]
requires = ["uv_build>=0.6.1,<0.11"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "PtyLab"
module-root = ""
